Solve some more warnings relating to conversion and some other stuff

+++ lib/my/src/my_revstr.c
@@ -10,9 +10,9 @@
 char *my_revstr(char *str)
 {
     char tmp;
-    int length = my_strlen(str);
+    size_t length = my_strlen(str);
 
-    for (int i = 0; i < (length / 2); ++i) {
+    for (size_t i = 0; i < (length / 2); ++i) {
         tmp = str[i];
         str[i] = str[length - i - 1];
         str[length - i - 1] = tmp;
